+++ crypto-primitives/src/crh/blake3/mod.rs
@@ -1,0 +1,2 @@
+mod crh;
+pub use crh::{Blake3CRH, Blake3TwoToOneCRH};