+++ crypto-primitives/src/crh/mod.rs
@@ -3,7 +3,11 @@ use crate::Error;
 use ark_serialize::{CanonicalDeserialize, CanonicalSerialize};
 use ark_std::{borrow::Borrow, fmt::Debug, hash::Hash, rand::Rng};
 
+#[cfg(feature = "blake3")]
+pub mod blake3;
 pub mod bowe_hopwood;
+pub mod byte_digest;
+pub use byte_digest::ByteDigest;
 #[cfg(feature = "constraints")]
 pub mod constraints;
 pub mod injective_map;