+++ crypto-primitives/src/merkle_tree/configs/mod.rs
@@ -1,0 +1,7 @@
+#[cfg(feature = "blake3")]
+mod blake3;
+mod poseidon;
+
+#[cfg(feature = "blake3")]
+pub use blake3::Blake3MerkleConfig;
+pub use poseidon::PoseidonMerkleConfig;