+++ crypto-primitives/src/merkle_tree/mod.rs
@@ -22,6 +22,8 @@ use rayon::prelude::*;
 #[cfg(feature = "constraints")]
 pub mod constraints;
 
+pub mod configs;
+
 #[cfg(test)]
 mod tests;